refactor: replace req.method handling to separate module
